Refactor(worklfow_manager): WorkflowManager, introducing WorkflowInstaller and WorkflowInvocationHandler classes for modularized tool installation, workflow upload handling, proper polling and invocation tracking, and cancellation
